Wikiracing (BS4, Docker, PostgreSQL)/requirements.txt
ratelimiter==1.2.0.post0
psycopg2-binary==2.9.5
beautifulsoup4==4.11.1
Wikipedia-API==0.5.8
requests==2.28.2
lxml==4.9.2
